+++ imports/api/attachments/attachments.js
@@ -8,13 +8,34 @@ export const Attachments = new FilesCollection({
   onBeforeUpload() {
     return true;
   },
-  onAfterUpload(file) {
+  onAfterUpload(fileRef) {
     if (Meteor.isServer) {
+      import { onAfterUpload as handleUpload } from "/imports/api/storage/server/upload";
+
       this.update(
-        { _id: file._id },
+        { _id: fileRef._id },
         { $set: { "meta.createdAt": new Date() } }
       );
+
+      handleUpload({
+        collection: Attachments,
+        folder: "attachments",
+        fileRef: fileRef
+      });
     }
+  },
+  interceptDownload(http, fileRef, version) {
+    if (Meteor.isServer) {
+      import { interceptDownload as handleDownload } from "/imports/api/storage/server/download";
+
+      return handleDownload({
+        collection: this,
+        http,
+        fileRef,
+        version
+      });
+    }
+    return false;
   }
 });
 
@@ -23,5 +44,16 @@ if (Meteor.isServer) {
     Attachments.collection.rawCollection().createIndex({ "meta.projectId": 1 });
     Attachments.collection.rawCollection().createIndex({ metataskId: 1 });
   });
+
+  // Intercept FilesCollection's remove method to remove file from other storages
+  const _origRemove = Attachments.remove;
+  Attachments.remove = function (search) {
+    import { onBeforeRemove } from "/imports/api/storage/server/remove";
+
+    onBeforeRemove({ collection: this.collection, search });
+
+    // remove original file from database
+    _origRemove.call(this, search);
+  };
 }
 Attachments.methods = {};